requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
dotenv==0.9.9
Flask==3.1.0
google-auth==2.38.0
google-genai==1.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pypdf==5.3.0
python-dotenv==1.0.1
reportlab==4.3.1
requests==2.32.3
rsa==4.9
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
